+++ atoms/src/main/java/com/vea/atoms/mvp/app/AppDelegate.java
@@ -21,7 +21,7 @@ import android.support.annotation.NonNull;
 
 import com.vea.atoms.mvp.di.component.AppComponent;
 import com.vea.atoms.mvp.di.component.DaggerAppComponent;
-import com.vea.atoms.mvp.di.modul.GlobalConfigModule;
+import com.vea.atoms.mvp.di.module.GlobalConfigModule;
 import com.vea.atoms.mvp.integration.ConfigModule;
 import com.vea.atoms.mvp.integration.ManifestParser;
 import com.vea.atoms.mvp.utils.AtomsUtils;
