+++ atoms/src/main/java/com/vea/atoms/mvp/di/module/AppModule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.vea.atoms.mvp.di.modul;
+package com.vea.atoms.mvp.di.module;
 
 
 import com.vea.atoms.mvp.integration.IRepositoryManager;
