+++ atoms/src/main/java/com/vea/atoms/mvp/http/log/FormatPrinter.java
@@ -16,7 +16,7 @@
 package com.vea.atoms.mvp.http.log;
 
 
-import com.vea.atoms.mvp.di.modul.GlobalConfigModule;
+import com.vea.atoms.mvp.di.module.GlobalConfigModule;
 
 import java.util.List;
