+++ atoms/src/main/java/com/vea/atoms/mvp/utils/AtomsUtils.java
@@ -17,13 +17,10 @@ package com.vea.atoms.mvp.utils;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.content.res.Resources;
 import android.graphics.drawable.Drawable;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.os.Message;
-import android.support.design.widget.Snackbar;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.RecyclerView;
 import android.text.SpannableString;
@@ -37,7 +34,6 @@ import android.view.WindowManager;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.vea.atoms.mvp.app.AppManager;
 import com.vea.atoms.mvp.app.IApp;
 import com.vea.atoms.mvp.di.component.AppComponent;
 
@@ -53,7 +49,6 @@ import java.security.MessageDigest;
  */
 public class AtomsUtils {
     static public Toast mToast;
-
 
     private AtomsUtils() {
         throw new IllegalStateException("you can't instantiate me!");
@@ -77,7 +72,6 @@ public class AtomsUtils {
         // 设置hint  
         v.setHint(new SpannedString(ss)); // 一定要进行转换,否则属性会消失
     }
-
 
     /**
      * dip转pix
@@ -112,7 +106,6 @@ public class AtomsUtils {
         return (int) (pix / densityDpi + 0.5f);
     }
 
-
     /**
      * 从 dimens 中获得尺寸
      *
@@ -140,7 +133,6 @@ public class AtomsUtils {
      *
      * @return
      */
-
     public static String getString(Context context, int stringID) {
         return getResources(context).getString(stringID);
     }
@@ -150,7 +142,6 @@ public class AtomsUtils {
      *
      * @return
      */
-
     public static String getString(Context context, String strName) {
         return getString(context, getResources(context).getIdentifier(strName, "string", context.getPackageName()));
     }
@@ -227,8 +218,6 @@ public class AtomsUtils {
         return getResources(context).getDrawable(rID);
     }
 
-
-
     /**
      * 获得屏幕的宽度
      *
@@ -243,10 +232,9 @@ public class AtomsUtils {
      *
      * @return
      */
-    public static int getScreenHeidth(Context context) {
+    public static int getScreenHeight(Context context) {
         return getResources(context).getDisplayMetrics().heightPixels;
     }
-
 
     /**
      * 获得颜色
@@ -282,7 +270,6 @@ public class AtomsUtils {
         return false;
     }
 
-
     /**
      * MD5
      *
@@ -308,9 +295,8 @@ public class AtomsUtils {
         return hex.toString();
     }
 
-
     /**
-     * 全屏,并且沉侵式状态栏
+     * 全屏,并且沉浸式状态栏
      *
      * @param activity
      */
@@ -320,23 +306,6 @@ public class AtomsUtils {
         activity.getWindow().setAttributes(attrs);
         activity.getWindow().addFlags(WindowManager.LayoutParams.FLAG_LAYOUT_IN_SCREEN);
         activity.getWindow().addFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
-    }
-
-
-    /**
-     * 配置 RecyclerView
-     *
-     * @param recyclerView
-     * @param layoutManager
-     * @deprecated Use {@link #configRecyclerView(RecyclerView, RecyclerView.LayoutManager)} instead
-     */
-    @Deprecated
-    public static void configRecycleView(final RecyclerView recyclerView
-            , RecyclerView.LayoutManager layoutManager) {
-        recyclerView.setLayoutManager(layoutManager);
-        //如果可以确定每个item的高度是固定的，设置这个选项可以提高性能
-        recyclerView.setHasFixedSize(true);
-        recyclerView.setItemAnimator(new DefaultItemAnimator());
     }
 
     /**
@@ -353,7 +322,6 @@ public class AtomsUtils {
         recyclerView.setItemAnimator(new DefaultItemAnimator());
     }
 
-
     public static AppComponent obtainAppComponentFromContext(Context context) {
         Preconditions.checkNotNull(context, "%s cannot be null", Context.class.getName());
         Preconditions.checkState(context.getApplicationContext() instanceof IApp, "Application does not implements IApp");
@@ -362,16 +330,17 @@ public class AtomsUtils {
 
     /**
      * 检查网络是否可用
+     *
      * @param context
-     * @return true,false
+     * @return true, false
      */
-    public static boolean isNetWorkAvailable(Context context){
+    public static boolean isNetWorkAvailable(Context context) {
 
         ConnectivityManager connectManager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
 
         NetworkInfo info = connectManager.getActiveNetworkInfo();
 
-        return (info!=null && info.isAvailable());
+        return (info != null && info.isAvailable());
 
     }
